Get @nteract/play into a working state
