--- README.md
@@ -1,31 +1,0 @@
-# @nteract/play
-
-Play is a simple REPL-like interface for executing code online. It allows users to write code snippets and execute their output. Play is powered by [Binder](https://mybinder.org/).
-
-## Installation
-
-To set up this project for development, start by cloning this repository.
-
-```
-$ git clone https://github.com/nteract/play.git
-```
-
-Install the dependencies for this project.
-
-```
-$ yarn
-```
-
-Then run the project in development mode.
-
-```
-$ yarn dev
-```
-
-## Support
-
-If you experience an issue while using this package or have a feature request, please file an issue on the [issue board](https://github.com/nteract/play/issues/new/choose).
-
-## License
-
-[BSD-3-Clause](https://choosealicense.com/licenses/bsd-3-clause/)
