+++ now.json
@@ -1,0 +1,3 @@
+{
+  "version": 2
+}
