+++ pages/_app.js
@@ -1,9 +1,10 @@
+
 import React from "react";
 import App, { Container } from "next/app";
 import withRedux from "next-redux-wrapper";
 import { Provider } from "react-redux";
 
-import { createStore } from "../app-redux";
+import { createStore } from "../redux";
 
 /**
  * @param {object} initialState
