+++ redux/actionTypes.js
@@ -1,3 +1,4 @@
+
 const PREFIX = "PLAY";
 const FULFILLED = "FULFILLED";
 const FAILED = "FAILED";
