+++ redux/createStore.js
@@ -7,27 +7,30 @@ import reducer from "./reducer";
 import getInitialState from "./getInitialState";
 import { errorMiddleware } from "./middlewares";
 
+/**
+ * Not that we need to here, but I thought I'd write it out
 type State = {
   ui: {
-    repo: string;
-    gitref: string;
-    source: string;
-    showPanel: boolean;
-    currentServerId: string;
-    currentKernelName: string;
-    platform: string;
-  };
+    repo: string,
+    gitref: string,
+    source: string,
+    showPanel: boolean,
+    currentServerId: string,
+    currentKernelName: string,
+    platform: string
+  },
   entities: {
-    serversById: any;
-  };
+    serversById: *
+  }
 };
+**/
 
-export default function(givenInitialState: Partial<State> = {}) {
+export default function(givenInitialState = {}) {
   const initialState = merge(getInitialState(), givenInitialState);
   const epicMiddleware = createEpicMiddleware();
   const composeEnhancers =
     (typeof window !== "undefined" &&
-      (window as any).__REDUX_DEVTOOLS_EXTENSION_COMPOSE__) ||
+      window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__) ||
     compose;
 
   const store = createStore(
