+++ redux/epics.js
@@ -1,21 +1,14 @@
 import { combineEpics, ofType } from "redux-observable";
 import { of, from, merge } from "rxjs";
 import { filter, catchError, mergeMap, switchMap } from "rxjs/operators";
-import { binder, IEventSource } from "rx-binder";
+import { binder } from "rx-binder";
 import { kernels, shutdown, kernelspecs } from "rx-jupyter";
 import uuid from "uuid/v4";
-import {
-  executeRequest,
-  kernelInfoRequest,
-  JupyterMessage,
-  MessageType
-} from "@nteract/messaging";
+import { executeRequest, kernelInfoRequest } from "@nteract/messaging";
 import objectPath from "object-path";
 
 import * as actions from "./actions";
 import * as actionTypes from "./actionTypes";
-
-type EventSource = new (url: string) => IEventSource;
 
 const activateServerEpic = action$ =>
   action$.pipe(
@@ -23,15 +16,11 @@ const activateServerEpic = action$ =>
     // Definitely do not run this on the server side
     filter(() => typeof window !== "undefined"),
     switchMap(({ payload: { serverId, oldServerId, repo, gitref } }) => {
-      return binder({ repo, ref: gitref }, EventSource).pipe(
+      return binder({ repo, gitref }, EventSource).pipe(
         mergeMap(message => {
-          const actionsArray: Array<{
-            type: string;
-            payload: {
-              serverId: string;
-              message?: string;
-            };
-          }> = [actions.addServerMessage({ serverId, message })];
+          const actionsArray = [
+            actions.addServerMessage({ serverId, message })
+          ];
           if (message.phase === "ready") {
             const config = {
               endpoint: message.url.replace(/\/\s*$/, ""),
@@ -117,14 +106,11 @@ const setActiveKernelEpic = (action$, state$) =>
         "channel"
       ];
       const channel = objectPath.get(state$.value, channelPath);
-      const actionsArray: Array<{
-        type: string;
-        payload: string | { serverId: string; kernelName: string };
-      }> = [actions.setCurrentKernelName(kernelName)];
+      const actionsArray = [actions.setCurrentKernelName(kernelName)];
       if (!channel) {
         actionsArray.push(actions.activateKernel({ serverId, kernelName }));
       }
-
+      // $FlowFixMe
       return of(...actionsArray);
     })
   );
@@ -151,16 +137,10 @@ const initializeKernelMessaging = action$ =>
       // Side effect! Get that Kernel Info!
       kernel.channel.next(kernelInfoRequest());
       return kernel.channel.pipe(
-        mergeMap((message: JupyterMessage<MessageType, any>) => {
-          const actionsArray: Array<{
-            type: string;
-            payload: {
-              serverId: string;
-              kernelName: string;
-              message?: JupyterMessage;
-              status?: string;
-            };
-          }> = [actions.addKernelMessage({ serverId, kernelName, message })];
+        mergeMap(message => {
+          const actionsArray = [
+            actions.addKernelMessage({ serverId, kernelName, message })
+          ];
           switch (message.header.msg_type) {
             case "status":
               actionsArray.push(
@@ -220,11 +200,7 @@ const activateKernelEpic = (action$, state$) =>
         mergeMap(data => {
           const session = uuid();
           const kernel = Object.assign({}, data.response, {
-            channel: kernels.connect(
-              config,
-              data.response.id,
-              session
-            )
+            channel: kernels.connect(config, data.response.id, session)
           });
 
           kernel.channel.next(kernelInfoRequest());
