+++ redux/getInitialState.js
@@ -1,3 +1,4 @@
+
 const getInitialState = () => ({
   ui: {
     repo: "binder-examples/jupyter-stacks",
