+++ redux/index.js
@@ -1,4 +1,5 @@
-import * as actions from "./actions";
+
+import * as actions from "./actions.js";
 
 export { actions };
 export { default as createStore } from "./createStore";
