+++ utils/createObjectReducer.js
@@ -1,5 +1,3 @@
-import { Action } from "redux";
-
 /**
  * Similar to how combineReducers allows you to create a container reducer to
  * delegate actions where each key maps to a reducer to handle the key's value--
@@ -11,14 +9,7 @@ import { Action } from "redux";
  * @param {Function} config.valueReducer A standard redux reducer function.
  * @returns {Function} A standard redux reducer function.
  */
-type Config = {
-  getKey(action: Action): string;
-  valueReducer: any;
-};
-const createObjectReducer = (config: Config) => (
-  state: any = {},
-  action: Action
-) => {
+const createObjectReducer = config => (state = {}, action) => {
   const key = config.getKey(action);
   if (typeof key !== "undefined") {
     const previousValueState = state[key];
