+++ utils/makeServerId.js
@@ -1,0 +1,3 @@
+import { formBinderURL } from "rx-binder";
+
+export default ({ repo, gitref }) => formBinderURL({ repo, gitref });
