--- utils/makeServerId.ts
@@ -1,4 +1,0 @@
-import { formBinderURL } from "rx-binder";
-
-export default ({ repo, gitref }: { repo: string; gitref: string }) =>
-  formBinderURL({ repo, ref: gitref });
